venkat/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
branca==0.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
diff-match-patch==20230430
Django==5.0.6
django-import-export==4.1.0
folium==0.17.0
frozenlist==1.4.1
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
migrations==0.1.1
multidict==6.1.0
mysql==0.0.3
mysqlclient==2.2.4
numpy==2.1.1
opencage==3.0.2
phonenumbers==8.13.46
pillow==10.4.0
requests==2.32.3
sqlparse==0.5.0
tablib==3.5.0
tqdm==4.66.5
tzdata==2024.1
urllib3==2.2.3
xyzservices==2024.9.0
yarl==1.13.1
